+++ src/components/date_picker/super_date_picker/super_date_picker.js
@@ -312,6 +312,7 @@ export class EuiSuperDatePicker extends Component {
           isCustom
           startDateControl={<div />}
           endDateControl={<div />}
+          onKeyPress={this.handleKeyPress}
           readOnly>
           <span className="euiSuperDatePicker__prettyFormat">
             {prettyInterval(this.props.isPaused, this.props.refreshInterval)}
@@ -412,6 +413,12 @@ export class EuiSuperDatePicker extends Component {
     }
   };
 
+  handleKeyPress = event => {
+    if (event.key === 'Enter') {
+      this.handleClickUpdateButton();
+    }
+  };
+
   renderUpdateButton = () => {
     if (!this.props.showUpdateButton || this.props.isAutoRefreshOnly) {
       return;
@@ -461,6 +468,7 @@ export class EuiSuperDatePicker extends Component {
       <EuiFlexGroup
         gutterSize="s"
         responsive={false}
+        onKeyPress={this.handleKeyPress}
         className={flexWrapperClasses}>
         <EuiFlexItem>
           <EuiFormControlLayout